[Task 3]
The classes BasePage, HomePage, PageManager, TabletsPage, AppleTabletsPage, AppleTabletsTest were refactored
