+++ src/main/java/ua/ek/base/BasePage.java
@@ -3,8 +3,6 @@ package ua.ek.base;
 import org.openqa.selenium.*;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.PageFactory;
-import org.openqa.selenium.support.ui.ExpectedConditions;
-import org.openqa.selenium.support.ui.WebDriverWait;
 import ua.ek.utils.WaitUtils;
 
 public abstract class BasePage {
@@ -97,7 +95,8 @@ public abstract class BasePage {
                 .click();
     }
 
-    protected void executeWebElement(WebElement webElement){
+    protected void clickElementWithJS(WebElement webElement){
+        waitUtils.elementToBeSelected(webElement, THREE_SECONDS);
         JavascriptExecutor js = (JavascriptExecutor) driver;
         js.executeScript("arguments[0].click();", webElement);
     }
