+++ src/test/java/ua/ek/tablets/manufacturers/AppleTabletsTest.java
@@ -1,9 +1,9 @@
-package ua.ek.xpath;
+package ua.ek.tablets.manufacturers;
 
 import org.testng.annotations.Test;
 import ua.ek.base.BaseTest;
 import ua.ek.pages.PageManager;
-import ua.ek.pages.xpath.AppleTabletsPage;
+import ua.ek.pages.tablets.manufacturers.AppleTabletsPage;
 import ua.ek.utils.AssertUtils;
 
 public class AppleTabletsTest extends BaseTest {
